requirements.txt
pandas==1.1.5
selenium==4.9.0
webdriver_manager==3.4.2
psycopg2--binary==2.8.6
azure-storage-blob==12.9.0
pytz==2022.1
chromedriver-autoinstaller==0.6.4
schedule==1.1.0
bs4==0.0.1


